Rework the project's metadata
The goal is to centralize all the metadata in a single place for other
scripts to easily retrieve it.

To keep the source code cleaner, the metadata could have been gathered
in a separate file but, since it can also be useful for the users to
programmatically query such metadata (mostly the `__version__`, really),
it is best to directly expose it from within the main module.